requirements.txt
Django==2.2
django-rest-framework==0.1.0
djangorestframework===3.11.0
Markdown==3.2.1
django-filter==2.2.0
whitenoise==5.0.1
djangoql===0.13.1
django-import-export===2.1.0
django-cors-headers==3.2.1